Refactor sheet music script into modules
