+++ sheet_music_pyqt5.py
@@ -43,809 +43,19 @@ from transposition_utils import (
 )
 from selenium import webdriver
 from selenium.webdriver.chrome.service import Service
-from selenium.webdriver.common.by import By
-from selenium.common.exceptions import (NoSuchElementException,
-                                        StaleElementReferenceException, TimeoutException)
-from selenium.webdriver.support.ui import WebDriverWait
-from selenium.webdriver.support import expected_conditions as EC
-
 from webdriver_manager.chrome import ChromeDriverManager
 import requests
-import torch
-import torch.nn as nn
-from torch.nn.functional import pad
-from reportlab.pdfgen import canvas
 
-# Local imports
-# Assuming model_functions.py is available and contains required functions
-from model_functions import UNet, VDSR, PIL_to_tensor, tensor_to_PIL, load_best_model
-
-
-# Check if CUDA is available
-cuda_available = torch.cuda.is_available()
-print(f"CUDA available: {cuda_available}")
-
-
-# XPath dictionary
-xpaths = {
-    'search_bar': '//*[@id="search-input-wrap"]/input',
-    'songs_parent': '//*[@id="page-wrapper"]/ion-router-outlet/app-page-search/ion-content/div/div/div/app-search/div',
-    'song_title': './div/a/div/h5',
-    'song_text3': './div/a/div/span/span',
-    'song_text2': './div/a/div/span',
-    'song_image': './div/div[1]/div/app-product-audio-preview-image/div/img',
-    'click_song': '//*[@id="page-wrapper"]/ion-router-outlet/app-page-search/ion-content/div/div/div/app-search/div/app-product-list-item[{index}]/div/a/div',
-    'chords_button': '//*[@id="page-wrapper"]/ion-router-outlet/app-product-page/ion-content/div/div/div[3]/div/div[1]/div[2]/div[1]/app-product-sheet-selector/div/div[1]/button',
-    'orchestration_header': "//h3[contains(text(), 'Orchestration')]/ancestor::div[4]",
-    'key_button': '//*[@id="page-wrapper"]/ion-router-outlet/app-product-page/ion-content/div/div/div[3]/div/div[1]/div[2]/div[1]/app-product-sheet-selector/div/div[3]/app-product-selector-key/div/button',
-    'key_parent': '//*[@id="page-wrapper"]/ion-router-outlet/app-product-page/ion-content/div/div/div[3]/div/div[1]/div[2]/div[1]/app-product-sheet-selector/div/div[3]/app-product-selector-key/div/ul',
-    'parts_button': '//*[@id="page-wrapper"]/ion-router-outlet/app-product-page/ion-content/div/div/div[3]/div/div[1]/div[2]/div[1]/app-product-sheet-selector/div/div[2]/div/button',
-    # 'parts_parent': '//*[@id="page-wrapper"]/ion-router-outlet/app-product-page/ion-content//ul',
-    'parts_parent': '//*[@id="page-wrapper"]/ion-router-outlet/app-product-page/ion-content/div/div/div[3]/div/div[1]/div[2]/div[1]/app-product-sheet-selector/div/div[2]/div/ul',
-    'parts_list': '//*[@id="page-wrapper"]/ion-router-outlet/app-product-page/ion-content//ul/li/button',
-    'image_element': '//*[@id="preview-sheets"]/div/div[1]/div/img',
-    'next_button': "//button[contains(@class, 'sheet-nav-gradient-button-right')]",
-}
-
-# Helper class for common Selenium operations
-class SeleniumHelper:
-    @staticmethod
-    def click_element(driver, xpath, timeout=2, log_func=None):
-        try:
-            element = WebDriverWait(driver, timeout).until(
-                EC.element_to_be_clickable((By.XPATH, xpath))
-            )
-            element.click()
-            return True
-        except (StaleElementReferenceException, NoSuchElementException, TimeoutException) as e:
-            if log_func:
-                log_func(f"Error clicking element at xpath: {xpath} - {str(e)}")
-            return False
-
-    @staticmethod
-    def find_element(driver, xpath, timeout=2, log_func=None):
-        try:
-            element = WebDriverWait(driver, timeout).until(
-                EC.presence_of_element_located((By.XPATH, xpath))
-            )
-            return element
-        except (StaleElementReferenceException, NoSuchElementException, TimeoutException) as e:
-            if log_func:
-                log_func(f"Error finding element at xpath: {xpath} - {str(e)}")
-            return None
-
-    @staticmethod
-    def find_elements(driver, xpath, timeout=2, log_func=None):
-        try:
-            elements = WebDriverWait(driver, timeout).until(
-                EC.presence_of_all_elements_located((By.XPATH, xpath))
-            )
-            return elements
-        except (StaleElementReferenceException, NoSuchElementException, TimeoutException) as e:
-            if log_func:
-                log_func(f"Error finding elements at xpath: {xpath} - {str(e)}")
-            return []
-
-    @staticmethod
-    def send_keys_to_element(driver, xpath, keys, timeout=2, log_func=None):
-        try:
-            element = WebDriverWait(driver, timeout).until(
-                EC.element_to_be_clickable((By.XPATH, xpath))
-            )
-            element.clear()
-            element.send_keys(keys)
-            return True
-        except (StaleElementReferenceException, NoSuchElementException, TimeoutException) as e:
-            if log_func:
-                log_func(f"Error sending keys to element at xpath: {xpath} - {str(e)}")
-            return False
-
-    @staticmethod
-    def click_dynamic_element(driver, xpath_template, index, timeout=2, log_func=None):
-        xpath = xpath_template.format(index=index)
-        return SeleniumHelper.click_element(driver, xpath, timeout=timeout, log_func=log_func)
-
-
-class FindSongsThread(QThread):
-    progress = pyqtSignal(int)
-    status = pyqtSignal(str)
-    song_info_updated = pyqtSignal(str, str)  # Updated to emit song text and image URL
-    song_choice_box_updated = pyqtSignal(str, str)  # Updated to emit song text and image URL
-    log_updated = pyqtSignal(str)
-    clear_song_info = pyqtSignal()
-    clear_song_choice_box = pyqtSignal()
-    clear_key_choice_box = pyqtSignal()
-    request_song_choice_box_count = pyqtSignal()
-    receive_song_choice_box_count = pyqtSignal(int)
-    insert_separator_in_song_choice_box = pyqtSignal(int)
-
-    def __init__(self, driver, user_song_choice):
-        super().__init__()
-        self.driver = driver
-        self.user_song_choice = user_song_choice
-        self.song_choice_box_count = 0
-        self.receive_song_choice_box_count.connect(self.set_song_choice_box_count)
-
-    def set_song_choice_box_count(self, count):
-        self.song_choice_box_count = count
-
-    def run(self):
-        self.status.emit("")
-        self.progress.emit(0)
-        self.clear_song_info.emit()
-        self.clear_song_choice_box.emit()
-        self.clear_key_choice_box.emit()
-
-        url = "https://www.praisecharts.com/search"
-        self.driver.get(url)
-
-        # Use SeleniumHelper to interact with elements
-        search_bar_xpath = xpaths['search_bar']
-        if not SeleniumHelper.send_keys_to_element(self.driver, search_bar_xpath, self.user_song_choice, log_func=self.log_updated.emit):
-            self.log_updated.emit("Error interacting with the search bar.")
-            return
-
-        songs_counter = 0
-
-        songs_parent_xpath = xpaths['songs_parent']
-        songs_parent = SeleniumHelper.find_element(self.driver, songs_parent_xpath, timeout=10, log_func=self.log_updated.emit)
-        if not songs_parent:
-            self.log_updated.emit("Error finding songs parent element.")
-            return
-
-        time.sleep(2)  # Ensure elements are loaded
-        songs_children = songs_parent.find_elements("xpath", './app-product-list-item')
-
-        for idx, child in enumerate(songs_children, 1):
-            title = ''
-            text2 = ''
-            text3 = ''
-            image_url = ''
-        
-            try:
-                title = child.find_element("xpath", xpaths['song_title']).text
-            except NoSuchElementException:
-                pass
-        
-            try:
-                text3 = child.find_element("xpath", xpaths['song_text3']).text
-            except NoSuchElementException:
-                pass
-        
-            if text3 != '':
-                try:
-                    text2 = child.find_element("xpath", xpaths['song_text2']).text.split("\n")[0]
-                except NoSuchElementException:
-                    pass
-        
-            if text3 == text2:
-                text2 = ''
-        
-            # Build element_text only from non-empty parts
-            element_text_parts = []
-            if title.strip():
-                element_text_parts.append(title.strip())
-            if text2.strip():
-                element_text_parts.append(text2.strip())
-            if text3.strip():
-                element_text_parts.append(text3.strip())
-            element_text = '\n'.join(element_text_parts)
-        
-            if not element_text:
-                continue  # Skip this iteration if element_text is empty
-        
-            try:
-                image_element = child.find_element("xpath", xpaths['song_image'])
-                image_url = image_element.get_attribute('src')
-            except NoSuchElementException:
-                image_url = ''
-        
-            if text3 != '':
-                self.song_info_updated.emit(element_text, image_url)
-                self.song_choice_box_updated.emit(element_text, image_url)
-                songs_counter += 1
-                self.request_song_choice_box_count.emit()
-            else:
-                self.song_info_updated.emit(title, image_url)
-
-
-        self.log_updated.emit(f"Found {songs_counter} songs for search: {self.user_song_choice}")
-
-
-class SelectSongThread(QThread):
-    progress = pyqtSignal(int)
-    status = pyqtSignal(str)
-    button_elements_signal = pyqtSignal(list)
-    log_updated = pyqtSignal(str)
-    clear_key_choice_box = pyqtSignal()
-    key_choice_box_updated = pyqtSignal(str)
-    clear_button_elements = pyqtSignal()
-    instrument_parts_signal = pyqtSignal(list)  # Signal to emit instrument parts
-    song_selection_failed = pyqtSignal()  # New signal to handle selection failures
-
-    def __init__(self, driver, selected_song, selected_song_index, selected_song_title, user_song_choice):
-        super().__init__()
-        self.driver = driver
-        self.selected_song = selected_song
-        self.user_song_choice = user_song_choice
-        self.selected_song_index = selected_song_index
-        self.selected_song_title = selected_song_title
-
-    def run(self):
-        try:
-            self.log_updated.emit(f"Selected song: {self.selected_song_title}")
-
-            # Click the selected song using SeleniumHelper
-            click_xpath_template = xpaths['click_song']
-            if not SeleniumHelper.click_dynamic_element(self.driver, click_xpath_template, self.selected_song_index + 1, log_func=self.log_updated.emit):
-                self.log_updated.emit("Error clicking the song.")
-                raise Exception("Error clicking the song.")
-
-            # Click "Chords & Lyrics" button
-            chords_click_xpath = xpaths['chords_button']
-            if not SeleniumHelper.click_element(self.driver, chords_click_xpath, log_func=self.log_updated.emit):
-                self.log_updated.emit("Error clicking 'Chords & Lyrics' button.")
-                raise Exception("Error clicking 'Chords & Lyrics' button.")
-
-            # Click "Orchestration" header
-            orch_click_xpath = xpaths['orchestration_header']
-            if not SeleniumHelper.click_element(self.driver, orch_click_xpath, log_func=self.log_updated.emit):
-                self.log_updated.emit("Orchestration not found for this song.")
-                self.song_selection_failed.emit()
-                return
-
-            # Access key menu
-            key_click_xpath = xpaths['key_button']
-            if not SeleniumHelper.click_element(self.driver, key_click_xpath, log_func=self.log_updated.emit):
-                self.log_updated.emit("Error accessing key menu.")
-                raise Exception("Error accessing key menu.")
-
-            key_parent_xpath = xpaths['key_parent']
-            key_parent_element = SeleniumHelper.find_element(self.driver, key_parent_xpath, log_func=self.log_updated.emit)
-            if not key_parent_element:
-                self.log_updated.emit("Error finding key parent element.")
-                raise Exception("Error finding key parent element.")
-
-            button_elements = key_parent_element.find_elements(by=By.TAG_NAME, value='button')
-            if not button_elements:
-                self.log_updated.emit("No key menu found.")
-                raise Exception("No key menu found.")
-
-            keys = []
-            self.clear_key_choice_box.emit()
-
-            for button in button_elements:
-                keys.append(button.text)
-                self.key_choice_box_updated.emit(button.text)
-
-            # Automatically select the first key
-            first_button = button_elements[0]
-            first_button.click()
-
-            formatted_keys = ', '.join(keys)
-            self.log_updated.emit(f"Found keys: {formatted_keys}")
-            self.log_updated.emit(f"Automatically selected key: {keys[0]}")
-            self.button_elements_signal.emit(button_elements)
-
-            # Call find_parts to get instrument parts
-            self.find_parts()
-
-        except Exception as e:
-            self.log_updated.emit(f"Exception during song selection: {str(e)}")
-            self.song_selection_failed.emit()
-
-
-
-    def find_parts(self):
-        # Click "Parts" button
-        parts_button_xpath = xpaths['parts_button']
-        if not SeleniumHelper.click_element(self.driver, parts_button_xpath, log_func=self.log_updated.emit):
-            self.log_updated.emit("Error accessing parts menu.")
-            return
-
-        # Find parts elements
-        parts_parent_xpath = xpaths['parts_parent']
-        parts_parent_element = SeleniumHelper.find_element(self.driver, parts_parent_xpath, log_func=self.log_updated.emit)
-        if not parts_parent_element:
-            self.log_updated.emit("Error finding parts parent element.")
-            return
-
-        parts_elements = parts_parent_element.find_elements(by=By.TAG_NAME, value='button')
-        if not parts_elements:
-            self.log_updated.emit("No parts menu found.")
-            return
-
-        instrument_parts = []
-
-        for part in parts_elements:
-            part_text = part.text.strip()
-            if not part_text:
-                continue  # Skip empty instrument names
-            if 'cover' not in part_text.lower() and 'lead sheet' not in part_text.lower():
-                instrument_parts.append(part_text)
-
-        # Close the parts menu
-        if not SeleniumHelper.click_element(self.driver, parts_button_xpath, log_func=self.log_updated.emit):
-            self.log_updated.emit("Error closing parts menu.")
-            return
-
-        self.instrument_parts_signal.emit(instrument_parts)
-
-
-
-class SelectKeyThread(QThread):
-    progress = pyqtSignal(int)
-    status = pyqtSignal(str)
-    log_updated = pyqtSignal(str)
-    instrument_parts_signal = pyqtSignal(list)  # Signal to emit instrument parts
-
-    def __init__(self, driver, selected_key, button_elements):
-        super().__init__()
-        self.driver = driver
-        self.selected_key = selected_key
-        self.button_elements = button_elements.copy()
-
-    def run(self):
-        # Click the key button
-        key_click_xpath = xpaths['key_button']
-        if not SeleniumHelper.click_element(self.driver, key_click_xpath, log_func=self.log_updated.emit):
-            pass
-
-        # Click the selected key
-        button_clicked = False
-
-        for button in self.button_elements:
-            if self.selected_key == button.text:
-                button.click()
-                button_clicked = True
-                break
-
-        if not button_clicked:
-            # Close the key menu if the key was not found
-            SeleniumHelper.click_element(self.driver, key_click_xpath, log_func=self.log_updated.emit)
-
-        self.log_updated.emit(f"Selected key: {self.selected_key}")
-
-        # Call find_parts to get instrument parts
-        self.find_parts()
-
-    def find_parts(self):
-        # Click "Parts" button
-        parts_button_xpath = xpaths['parts_button']
-        if not SeleniumHelper.click_element(self.driver, parts_button_xpath, log_func=self.log_updated.emit):
-            self.log_updated.emit("Error accessing parts menu.")
-            return
-
-        # Find parts elements
-        parts_list_xpath = xpaths['parts_list']
-        parts_elements = SeleniumHelper.find_elements(self.driver, parts_list_xpath, log_func=self.log_updated.emit)
-        if not parts_elements:
-            self.log_updated.emit("No parts menu found.")
-            return
-
-        instrument_parts = []
-
-        for part in parts_elements:
-            part_text = part.text.lower()
-            if 'cover' not in part_text and 'lead sheet' not in part_text:
-                instrument_parts.append(part.text)
-
-        # Close the parts menu
-        if not SeleniumHelper.click_element(self.driver, parts_button_xpath, log_func=self.log_updated.emit):
-            self.log_updated.emit("Error closing parts menu.")
-            return
-
-        self.instrument_parts_signal.emit(instrument_parts)
-
-
-class DownloadAndProcessThread(QThread):
-    progress = pyqtSignal(int)
-    status = pyqtSignal(str)
-    log_updated = pyqtSignal(str)
-
-    def __init__(self, driver, key_choice_text, selected_song_title, selected_song_artist, paths, selected_instruments, download_horn_only=False):
-        super().__init__()
-        self.driver = driver
-        self.key_choice_text = key_choice_text
-        self.selected_song_title = selected_song_title
-        self.selected_song_artist = selected_song_artist
-        self.paths = paths
-        self.download_horn_only = download_horn_only
-        self.selected_instruments = selected_instruments
-        self.instrument_parts = []  # Initialize here for debugging
-        self.full_paths = []  # Initialize here for debugging
-        self.images_by_instrument = defaultdict(list)  # Initialize here for debugging
-        self.log_updated.emit("Initialized DownloadAndProcessThread")
-
-    def run(self):
-        self.log_updated.emit("Starting run method")
-        try:
-            song_dir, temp_dir = self.initialize_directories()
-            self.log_updated.emit(f"Directories initialized: song_dir={song_dir}, temp_dir={temp_dir}")
-            
-            self.find_parts()
-            self.download_images(temp_dir)
-            self.remove_watermarks()
-            self.upscale_images()
-            
-            torch.cuda.empty_cache()
-            self.create_pdfs(song_dir, temp_dir)
-            self.cleanup(temp_dir)
-            self.open_directory(song_dir)
-        except Exception as e:
-            self.log_updated.emit(f"Exception in run: {str(e)}")
-
-    def initialize_directories(self):
-        self.log_updated.emit("Initializing directories")
-        try:
-            key_dir = self.key_choice_text
-            title_dir = re.sub(r'[<>:"\\|?* ]', '_', self.selected_song_title.replace("/", "-"))
-            artist_dir = re.sub(r'[<>:"\\|?* ]', '_', self.selected_song_artist.replace("/", "-"))
-            main_dir = self.paths['download_dir']
-            song_dir = os.path.join(main_dir, title_dir, artist_dir, key_dir)
-            os.makedirs(song_dir, exist_ok=True)
-            temp_dir = os.path.join(main_dir, title_dir, artist_dir, self.paths['temp_sub_dir'])
-            os.makedirs(temp_dir, exist_ok=True)
-            self.log_updated.emit(f"Directories created: song_dir={song_dir}, temp_dir={temp_dir}")
-            return song_dir, temp_dir
-        except Exception as e:
-            self.log_updated.emit(f"Error in initialize_directories: {str(e)}")
-            raise
-
-    def find_parts(self):
-        self.log_updated.emit("Finding parts")
-        # Click "Parts" button
-        parts_button_xpath = xpaths['parts_button']
-        if not SeleniumHelper.click_element(self.driver, parts_button_xpath, log_func=self.log_updated.emit):
-            self.log_updated.emit("Error accessing parts menu.")
-            return
-
-        # Find parts elements
-        parts_parent_xpath = xpaths['parts_parent']
-        parts_parent_element = SeleniumHelper.find_element(self.driver, parts_parent_xpath, log_func=self.log_updated.emit)
-        if not parts_parent_element:
-            self.log_updated.emit("Error finding parts parent element.")
-            return
-
-        parts_elements = parts_parent_element.find_elements(by=By.TAG_NAME, value='button')
-        if not parts_elements:
-            self.log_updated.emit("No parts menu found.")
-            return
-
-        self.instrument_parts = []
-
-        for part in parts_elements:
-            part_text = part.text.lower()
-            if 'cover' not in part_text and 'lead sheet' not in part_text:
-                self.instrument_parts.append(part.text)
-
-        # Close the parts menu
-        if not SeleniumHelper.click_element(self.driver, parts_button_xpath, log_func=self.log_updated.emit):
-            self.log_updated.emit("Error closing parts menu.")
-            return
-
-        self.log_updated.emit(f"Instrument parts found: {self.instrument_parts}")
-
-    def download_images(self, temp_dir):
-        self.log_updated.emit("Starting download_images")
-
-        # Define variables to track downloaded URLs and paths
-        self.images_by_instrument = defaultdict(list)
-        downloaded_urls = set()
-
-        try:
-            image_xpath = xpaths['image_element']
-            next_button_xpath = xpaths['next_button']
-            parts_button_xpath = xpaths['parts_button']
-            parts_list_xpath = xpaths['parts_list']
-
-            # Regex pattern to extract the page number (last 3 digits before .png)
-            page_number_pattern = re.compile(r'_(\d{3})\.png$')
-
-            instruments_to_process = []
-
-            if self.download_horn_only:
-                self.log_updated.emit("French horn only mode enabled.")
-                # Directly select "French Horn" from the parts
-                horn_instrument = None
-                for part in self.instrument_parts:
-                    if 'french horn' in part.lower():
-                        horn_instrument = part
-                        instruments_to_process.append(part)
-                        self.log_updated.emit(f"Selected instrument for French horn only: {part}")
-                        break
-                if not horn_instrument:
-                    self.log_updated.emit("No instrument with 'French Horn' found.")
-            elif self.selected_instruments:
-                instruments_to_process = self.selected_instruments
-                self.log_updated.emit(f"Selected instruments to process: {instruments_to_process}")
-            else:
-                # No instruments selected; process all parts
-                instruments_to_process = self.instrument_parts
-                self.log_updated.emit(f"No instruments selected. Processing all parts: {instruments_to_process}")
-
-            for instrument in instruments_to_process:
-                self.log_updated.emit(f"Processing instrument: {instrument}")
-
-                # Click the "Parts" button
-                if not SeleniumHelper.click_element(self.driver, parts_button_xpath, log_func=self.log_updated.emit):
-                    self.log_updated.emit('Error clicking "Parts" button')
-                    continue
-                self.log_updated.emit('Clicked "Parts" button')
-
-                # Find and click the instrument in the dropdown list
-                instrument_list_elements = SeleniumHelper.find_elements(self.driver, parts_list_xpath, log_func=self.log_updated.emit)
-                if not instrument_list_elements:
-                    self.log_updated.emit("Error finding instrument list elements.")
-                    continue
-
-                instrument_clicked = False
-                for button in instrument_list_elements:
-                    button_text = button.text.strip()
-                    normalized_instrument = instrument.lower().replace(",", "").replace("-", " ").strip()
-                    normalized_button_text = button_text.lower().replace(",", "").replace("-", " ").strip()
-
-                    if normalized_instrument == normalized_button_text:
-                        try:
-                            button.click()
-                            instrument_clicked = True
-                            self.log_updated.emit(f"Selected instrument: {instrument}")
-                            break
-                        except Exception as e:
-                            self.log_updated.emit(f"Error clicking instrument '{instrument}': {str(e)}")
-                            continue
-
-                if not instrument_clicked:
-                    self.log_updated.emit(f"Could not find instrument: {instrument} in the dropdown list.")
-                    continue
-
-                # Download images for the selected instrument
-                previous_page_number = None
-
-                while True:
-                    # Locate the image element
-                    image_element = SeleniumHelper.find_element(self.driver, image_xpath, log_func=self.log_updated.emit)
-                    if not image_element:
-                        self.log_updated.emit(f"No more images found for {instrument}.")
-                        break
-
-                    image_url = image_element.get_attribute('src')
-
-                    # Avoid downloading duplicate URLs
-                    if image_url in downloaded_urls:
-                        self.log_updated.emit(f"URL already downloaded: {image_url}, skipping.")
-                        if not self.click_next_button(next_button_xpath):
-                            break
-                        continue
-
-                    # Extract the page number
-                    match = page_number_pattern.search(image_url)
-                    if match:
-                        current_page_number = match.group(1)
-                    else:
-                        self.log_updated.emit(f"Could not extract page number from URL: {image_url}")
-                        break
-
-                    # If the page number resets to "001", stop downloading for this instrument
-                    if previous_page_number and current_page_number == "001":
-                        self.log_updated.emit(f"Page number reset to '001'. Stopping download for {instrument}.")
-                        break
-
-                    # Download the image
-                    self.status.emit(f"Downloading {os.path.basename(image_url)}")
-                    full_path = os.path.join(temp_dir, os.path.basename(image_url))
-                    downloaded_urls.add(image_url)
-
-                    try:
-                        response = requests.get(image_url)
-                        if response.status_code == 200:
-                            with open(full_path, 'wb') as f:
-                                f.write(response.content)
-                            self.log_updated.emit(f"Downloaded image: {os.path.basename(image_url)}")
-                            self.images_by_instrument[instrument].append(full_path)
-                        else:
-                            self.log_updated.emit(f"Failed to download image: {os.path.basename(image_url)} - Status code: {response.status_code}")
-
-                        # Click the next button to move to the next image
-                        if not self.click_next_button(next_button_xpath):
-                            self.log_updated.emit(f"No next button available. Stopping download for {instrument}.")
-                            break
-
-                    except Exception as e:
-                        self.log_updated.emit(f"An error occurred while downloading images: {e}")
-                        break
-
-                    # Update the previous page number
-                    previous_page_number = current_page_number
-
-        except Exception as e:
-            self.log_updated.emit(f"Exception in download_images: {str(e)}")
-
-    def click_next_button(self, next_button_xpath):
-        """Clicks the 'Next' button to navigate through images."""
-        self.log_updated.emit("Attempting to click next button")
-        if SeleniumHelper.click_element(self.driver, next_button_xpath, log_func=self.log_updated.emit):
-            self.log_updated.emit("Clicked next button")
-            return True
-        else:
-            self.log_updated.emit("Error clicking next button or no next button available.")
-            return False
-        return False
-
-    def remove_watermarks(self):
-        self.log_updated.emit("Starting remove_watermarks")
-        try:
-            self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-            wm_model = UNet().to(self.device)
-            load_best_model(wm_model, self.paths['wm_model_path'])
-            wm_model.eval()
-
-            self.wm_outputs = defaultdict(list)
-            total_images = sum(len(paths) for paths in self.images_by_instrument.values())
-            processed_images = 0
-
-            self.status.emit("Removing watermarks")
-            self.progress.emit(0)
-
-            with torch.inference_mode():
-                for instrument, paths in self.images_by_instrument.items():
-                    for path in paths:
-                        try:
-                            image_tensor = PIL_to_tensor(path).unsqueeze(0).to(self.device)
-                            wm_output = wm_model(image_tensor)
-                            self.wm_outputs[instrument].append(wm_output.cpu())
-                            self.status.emit(f"Unwatermarking {os.path.basename(path)}")
-                            processed_images += 1
-                            progress_value = int((processed_images / total_images) * 100)
-                            self.progress.emit(progress_value)
-                        except Exception as e:
-                            self.log_updated.emit(f"Error unwatermarking {path}: {str(e)}")
-
-            self.log_updated.emit("Removed watermarks")
-        except Exception as e:
-            self.log_updated.emit(f"Exception in remove_watermarks: {str(e)}")
-
-    def upscale_images(self):
-        self.log_updated.emit("Starting upscale_images")
-        try:
-            image_base_width, image_base_height = 1700, 2200
-
-            us_model = VDSR().to(self.device)
-            load_best_model(us_model, self.paths['us_model_path'])
-            us_model.eval()
-
-            upsample = nn.Upsample(size=(image_base_height, image_base_width), mode='nearest')
-
-            self.us_outputs = defaultdict(list)
-            total_images = sum(len(outputs) for outputs in self.wm_outputs.values())
-            processed_images = 0
-
-            self.status.emit("Upscaling images")
-            self.progress.emit(0)
-
-            with torch.inference_mode():
-                for instrument, wm_outputs in self.wm_outputs.items():
-                    for wm_output in wm_outputs:
-                        try:
-                            # Upsample the image
-                            wm_output_upscaled = upsample(wm_output)
-                            padding_size = 16
-                            patch_height = 550
-                            patch_width = 850
-
-                            padding = (padding_size, padding_size, padding_size, padding_size)
-                            wm_output_upscaled_padded = nn.functional.pad(wm_output_upscaled, padding, value=1.0)
-                            us_output = torch.zeros_like(wm_output_upscaled).cpu()
-
-                            # Patch processing
-                            for i in range(0, wm_output_upscaled.shape[-2], patch_height):
-                                for j in range(0, wm_output_upscaled.shape[-1], patch_width):
-                                    patch = wm_output_upscaled_padded[:, :, i:i + patch_height + padding_size * 2, j:j + patch_width + padding_size * 2].cpu()
-                                    us_patch = us_model(patch.to(self.device))
-                                    us_patch = us_patch[:, :, padding_size:-padding_size, padding_size:-padding_size]
-                                    us_output[:, :, i:i + patch_height, j:j + patch_width] = us_patch.cpu()
-
-                            self.us_outputs[instrument].append(us_output)
-                            self.status.emit(f"Upscaled image for {instrument}")
-                            processed_images += 1
-                            progress_value = int((processed_images / total_images) * 100)
-                            self.progress.emit(progress_value)
-                        except Exception as e:
-                            self.log_updated.emit(f"Error upscaling image for {instrument}: {str(e)}")
-
-            self.log_updated.emit("Upscaled images")
-        except Exception as e:
-            self.log_updated.emit(f"Exception in upscale_images: {str(e)}")
-
-    def create_pdfs(self, song_dir, temp_dir):
-        self.log_updated.emit("Starting create_pdfs")
-        try:
-            img_width, img_height = 1700, 2200
-
-            total_instruments = len(self.us_outputs)
-            processed_instruments = 0
-
-            for instrument, us_outputs in self.us_outputs.items():
-                # Get the original filename from the first image path
-                image_paths = self.images_by_instrument[instrument]
-                first_image_path = image_paths[0]
-                first_image_filename = os.path.basename(first_image_path)
-                # Remove the page number from the filename
-                base_filename = re.sub(r'_\d{3}\.png$', '', first_image_filename)
-                # Also remove any extension
-                base_filename = os.path.splitext(base_filename)[0]
-                pdf_filename = f"{base_filename}.pdf"
-                pdf_path = os.path.join(song_dir, pdf_filename)
-                c = canvas.Canvas(pdf_path, pagesize=(img_width, img_height))
-                self.status.emit(f"Creating {pdf_filename}")
-
-                for idx, image_tensor in enumerate(us_outputs):
-                    try:
-                        # Convert tensor to PIL image
-                        image_pil = tensor_to_PIL(image_tensor.squeeze(0))  # Remove batch dimension
-
-                        # Create a sanitized filename for the temporary image
-                        temp_image_name = f"temp_image_{base_filename}_{idx}.png"
-                        temp_image_path = os.path.join(temp_dir, temp_image_name)
-                        image_pil.save(temp_image_path)
-
-                        # Draw image on PDF
-                        c.drawImage(temp_image_path, 0, 0, width=img_width, height=img_height)
-                        c.showPage()
-
-                        # Remove temporary image
-                        os.remove(temp_image_path)
-                    except Exception as e:
-                        self.log_updated.emit(f"Error creating PDF for {instrument}, image {idx}: {str(e)}")
-
-                c.save()
-                self.log_updated.emit(f"Created PDF: {pdf_path}")
-                processed_instruments += 1
-                progress_value = int((processed_instruments / total_instruments) * 100)
-                self.progress.emit(progress_value)
-
-            self.status.emit(f"Process completed for {self.selected_song_title}")
-            self.progress.emit(0)
-            self.log_updated.emit("Processed images and created PDFs")
-        except Exception as e:
-            self.log_updated.emit(f"Exception in create_pdfs: {str(e)}")
-        
-    def cleanup(self, temp_dir):
-        self.log_updated.emit("Starting cleanup")
-        try:
-            for paths in self.images_by_instrument.values():
-                for path in paths:
-                    os.remove(path)
-                    self.log_updated.emit(f"Removed file: {path}")
-            os.rmdir(temp_dir)
-            self.log_updated.emit(f"Removed temp directory: {temp_dir}")
-            self.log_updated.emit(f"Process completed for song: {self.selected_song_title}")
-        except Exception as e:
-            self.log_updated.emit(f"Exception in cleanup: {str(e)}")
-
-    def open_directory(self, path):
-        self.log_updated.emit(f"Attempting to open directory: {path}")
-        try:
-            if not os.environ.get('DISPLAY') and platform.system() == 'Linux':
-                self.log_updated.emit("No display server available. Skipping open directory.")
-                return
-
-            if platform.system() == "Windows":
-                # Redirect stdout and stderr to suppress errors
-                subprocess.run(['explorer', path.replace('/', '\\')], stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-            elif platform.system() == "Darwin":
-                # Redirect stdout and stderr to suppress errors
-                subprocess.run(['open', path], stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-            elif platform.system() == "Linux":
-                # Redirect stdout and stderr to suppress errors
-                subprocess.run(['xdg-open', path], stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-            else:
-                self.log_updated.emit(f"Unsupported platform: {platform.system()}")
-        except Exception as e:
-            self.log_updated.emit(f"Exception in open_directory: {str(e)}")
-            self.log_updated.emit(f"Unsupported platform: {platform.system()}")
-
-
+# Imports moved to dedicated modules
+from selenium_utils import SeleniumHelper, xpaths
+from sheet_music_threads import (
+    FindSongsThread,
+    SelectSongThread,
+    SelectKeyThread,
+    DownloadAndProcessThread,
+)
+
+# Main application window
 class App(QMainWindow):
     def __init__(self, *args, **kwargs):
         super(App, self).__init__(*args, **kwargs)
